+++ lib/posts.js
@@ -6,18 +6,9 @@ import html from 'remark-html';
 import prism from 'remark-prism';
 import emoji from 'remark-emoji';
 import readingTime from 'reading-time';
+import { formatDate } from './utils/formatDate';
 
 const postsDirectory = path.join(process.cwd(), '/articles');
-
-const formatDate = (date) => {
-  const options = {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-  };
-
-  return new Intl.DateTimeFormat('en-US', options).format(new Date(date));
-};
 
 export function getSortedPostsData() {
   const fileNames = fs.readdirSync(postsDirectory);
@@ -71,6 +62,7 @@ export async function getPostData(id) {
   const { data, content } = matter(fileContents);
   const stats = readingTime(content);
   const formattedDate = formatDate(data.date);
+  const tags = data.tag.split(',');
 
   const processedContent = await remark()
     .use(html, { sanitize: false })
@@ -83,6 +75,7 @@ export async function getPostData(id) {
     id,
     content: contentHtml,
     data,
+    tags,
     stats,
     formattedDate,
   };
