+++ lib/utils/formatDate.js
@@ -1,0 +1,11 @@
+const formatDate = (date) => {
+  const options = {
+    year: 'numeric',
+    month: 'short',
+    day: 'numeric',
+  };
+
+  return new Intl.DateTimeFormat('en-US', options).format(new Date(date));
+};
+
+export { formatDate };
